+++ src/core/planting-spots/application/use-cases/get-planting-spot/get-planting-spot.use-case.spec.ts
@@ -10,6 +10,7 @@ const mockSpot: PlantingSpot = {
   description: null,
   userId: 'user-1',
   spaceId: 'space-1',
+  resolvedPlants: [],
   createdAt: '2024-01-01',
   updatedAt: '2024-01-01',
 };