+++ src/core/planting-spots/application/use-cases/update-planting-spot/update-planting-spot.use-case.spec.ts
@@ -11,6 +11,7 @@ const mockSpot: PlantingSpot = {
   description: null,
   userId: 'user-1',
   spaceId: 'space-1',
+  resolvedPlants: [],
   createdAt: '2024-01-01',
   updatedAt: '2024-01-02',
 };