+++ src/core/planting-spots/infrastructure/repositories/graphql/queries/planting-spots-find-by-criteria.query.ts
@@ -1,18 +1,39 @@
 import { gql } from '@apollo/client';
 
 export const PLANTING_SPOTS_FIND_BY_CRITERIA = gql`
-  query PlantingSpotsFindByCriteria {
-    plantingSpotsFindByCriteria {
+  query PlantingSpotsFindByCriteria($input: PlantingSpotFindByCriteriaRequestDto) {
+    plantingSpotsFindByCriteria(input: $input) {
       items {
         id
         name
         type
         description
+        capacity
+        row
+        column
+        dimensionsWidth
+        dimensionsHeight
+        dimensionsLength
+        soilType
         userId
         spaceId
+        resolvedPlants {
+          id
+          name
+          plantSpeciesId
+          imageUrl
+          userId
+          spaceId
+          createdAt
+          updatedAt
+        }
         createdAt
         updatedAt
       }
+      total
+      page
+      perPage
+      totalPages
     }
   }
 `;