+++ src/core/spaces/domain/interfaces/space-weather.interface.ts
@@ -1,0 +1,8 @@
+import type { DailyForecast } from '@/core/spaces/domain/interfaces/daily-forecast.interface';
+
+export interface SpaceWeather {
+  latitude: number;
+  longitude: number;
+  timezone: string;
+  daily: DailyForecast[];
+}